refactor: Split up codebase
