+++ src/lib.rs
@@ -1,7 +1,13 @@
 extern crate serde;
 extern crate serde_json;
 
-mod error;
-pub use error::*;
 mod cargo;
 pub use cargo::*;
+mod build;
+pub use build::*;
+mod test;
+pub use test::*;
+mod msg;
+pub use msg::*;
+mod error;
+pub use error::*;
